+++ generic/tclInt.h
@@ -2895,8 +2895,8 @@ MODULE_SCOPE void	TclArgumentGet(Tcl_Interp *interp, Tcl_Obj *obj,
 			    CmdFrame **cfPtrPtr, int *wordPtr);
 MODULE_SCOPE int	TclArraySet(Tcl_Interp *interp,
 			    Tcl_Obj *arrayNameObj, Tcl_Obj *arrayElemObj);
-MODULE_SCOPE void	TclBAConvertIndices(size_t index, size_t *hiPtr,
-			    size_t *loPtr);
+MODULE_SCOPE void	TclBAConvertIndices(size_t index, unsigned int *hiPtr,
+			    unsigned int *loPtr);
 MODULE_SCOPE double	TclBignumToDouble(const mp_int *bignum);
 MODULE_SCOPE int	TclByteArrayMatch(const unsigned char *string,
 			    int strLen, const unsigned char *pattern,
